refactor : 모듈화
