+++ src/constants/constants.js
@@ -1,0 +1,5 @@
+// 비교 할 수의 총 갯수를 지정하는 상수
+
+const COMPARE_LENGTH = 3;
+
+export default COMPARE_LENGTH;
