+++ src/constants/messages.js
@@ -1,0 +1,13 @@
+// 출력 할 때 사용할 메시지를 지정하는 상수
+
+const MESSAGE = {
+  START: "숫자 야구 게임을 시작합니다.",
+  INPUT: "숫자를 입력해주세요 : ",
+  RESTART: "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.\n",
+  SUCCESS: "개의 숫자를 모두 맞히셨습니다! 게임 종료",
+  STRIKE: "스트라이크",
+  BALL: "볼",
+  NOTHING: "낫싱",
+};
+
+export default MESSAGE;
